fixed for high numbers

+++ src/nchooser/NChooseR.java
@@ -21,10 +21,10 @@ public class NChooseR {
         //
         int chosen = Integer.parseInt(JOptionPane.showInputDialog("How many objects to choose ?"));
         int objects = Integer.parseInt(JOptionPane.showInputDialog("How many objects are there to choose from ?"));
-        int answer = factorial(objects)/(factorial(chosen)*factorial(objects - chosen));
+        long answer = factorial(objects)/(factorial(chosen)*factorial(objects - chosen));
         JOptionPane.showMessageDialog(null,"There are "+answer+" ways to choose "+chosen+" from a set of "+objects+" objects.");
     }
-    public static int factorial(int n){
+    public static long factorial(int n){
         //Factorial calculation
         if((n == 0) || (n == 1)){
             return 1;
